Link `*san_with_fuzzer.so` with `-ldl`
The `*san_with_fuzzer.so` dynamic libraries include libFuzzer, which requires `libdl.so`. However, in `merge_libfuzzer_sanitizer.sh`, the libraries are not linked with `-ldl`, which means that they can only be preloaded into a binary that itself links in `libdl.so`. While `python` itself does this, shells might not, which can lead to errors when the preload is applied to e.g. a wrapper script that unpacks a packaged Python application and only then invokes `python` itself.

This is fixed by linking `*san_with_fuzzer.so` with `-ldl`.