Remove references to Hypothesis
The Hypothesis documentation and examples were contributed externally, but
recent testing has shown that it performs quite poorly. Hypothesis severs the
connection between comparison coverage events and the fuzz input.

This might work better if implemented as a custom libFuzzer mutator. For now,
delete the examples from the repo to prevent confusion.

PiperOrigin-RevId: 385800165
Change-Id: I96693809e0083c2ad2c1bc719a96d50e2f9ad1e3