Remove calls that are not async-signal-safe from SIGALRM handler
Atheris will now attempt to use a Python signal handler for reporting Python backtraces. This does not always succeed because execution must return to Python code before that handler can be executed; if it fails to happen within 1 second, fall back to the libFuzzer handler, which prints native traces only.

PiperOrigin-RevId: 388670027
Change-Id: I8e016a419815de480a456fe5acf390cd407dee0a