Use multiple strategies for instrumenting imports
The previous change fixed the crash caused by the previous version of this feature on certain custom importers, but didn't actually work for some other custom importers.

This change now uses 2 different strategies for custom importers: it tries regular inheritance (if the importer is a class); or object inheritance (if the importer is an object). If object inheritance fails (typically because the importer is a native extension), instrumentation is completely skipped.

This also adds a feature to completely turn off instrumentation for modules imported via custom importers.

Finally, this adds tests.

PiperOrigin-RevId: 390727264
Change-Id: I5dc60061a11cb255b85d98a13cc1b8f1b05ed593