Avoid exception conversion bug in pybind11 by formatting exceptions ourselves
If the message in a Python exception contains surrogate characters, pybind11 fails to correctly translate it.

In older versions, this presents as Pybind11 replacing the original exception with RuntimeError: Unable to extract string contents! (encoding issue). This is at least descriptive, but it completely replaces the original exception with that message, making it impossible to determine the original exception. Furthermore, the replacement happened during the Python->C++ transition, so there's no way to get the original exception before it's replaced either. (Whether this behavior in the older version of pybind11 is "buggy" is ambiguous).

In new versions of pybind11, however, the cast_error is thrown from within the .what() function on pybind11::error_already_set. Unfortunately, .what() is marked as noexcept, meaning it promises never to throw an exception. This causes Python to crash without any opportunity to handle the error.

https://github.com/google/oss-fuzz/issues/8682

PiperOrigin-RevId: 483509304
Change-Id: Ic258af8b718c9a1e1ee7f7fca9a0e983eedf559a