Fixes the str method call instruction replacement issue when branching is involved.
The newly released Atheris feature to support tracing str methods (startswith and endswith) sometimes failed when a jump instruction was in the bytecode. This occurred because a pseudo stack is used to determine if certain call instructions should be replaced. When branch/jump instructions are involved, the pseudo stack does not get updated perfectly since the pseudo stack is affected by all the branches (when the real stack during runtime would only be affected by one branch). Because of this, some call instructions that weren't supposed to be replaced were being replaced, and this was creating faulty bytecode.

The solution was to remove the method from the traced_methods list once the method's corresponding call instruction was replaced.

PiperOrigin-RevId: 554867851
Change-Id: I04acf9d2bf430ab3091ecd7981b07d999cea49e0