Fix `.startswith` and `.endswith` support
Since 3.11, calls can be made with the `CALL` instruction in addition to other more specialized instructions. Furthermore, the argument to load_attr is left-shifted by 1, and the function call to normalize it was missing in one place, causing us to attempt to instrument the wrong thing (often our own instrumentation call) rather than the string functions.

PiperOrigin-RevId: 810958038
Change-Id: I0b91bb2f93934f8227d2f1ed7015696208d2a9c0