Rewrite instrument_bytecode.py to use references rather than offsets
instrument_bytecode.py spends a lot of effort trying to keep offsets correct after inserting instructions. The code to do so is extremely bug-prone. This change completely replaces how that works, with a technique that's much simpler and easy to reason about.

It now happens in the following steps:
 1. When code is parsed, it's no longer broken into basic blocks (though we do determine the *start* of every basic block, as instructions that are the destination of at least one jump)
 2. During parsing, each jump instruction's destination instruction is calculated; then the offsets are thrown away, and a reference to the destination instruction is kept instead.
 3. EXTENDED_ARG and CACHE instructions are combined into a single logical "Instruction" at parse-time.
 4. Instrumentation performs its transformations on this simplified series of instructions without worrying about offsets, CACHEs, or EXTENDED_ARGs.
 5. After all instrumentation is performed, new offsets and EXTENDED_ARGs are calculated, and the code is emitted.

This change adds a new `clean_instrument_bytecode` module for 3.12+; the old `instrument_bytecode` module is kept unchanged. This is temporary; I will either ensure `clean_instrument_bytecode` can handle older Python versions, or get rid of `instrument_bytecode` and only support python 3.12+.

This change does not include support for `str.startswith` and `str.endswith`. I have support for that mostly complete at the time of this writing, but I will include it in another CL.

PiperOrigin-RevId: 819329880
Change-Id: I58923cdde48dbb304c6dfd1eac89f954d67705e1