Add mock libFuzzer build mode and improve tests.
This mock libFuzzer exposes hooks allowing each of the libFuzzer sanitizer callbacks to be mocked in C++. A Python shim then allows those hooks to be accessed from Python. A new str_hook_test is added to test the string/bytes hooking mechanisms, which required this feature; and instrument_bytecode_test is updated to use this new feature as well.

This change found two bugs, which it fixes:
 - The str/bytes hooking mechanism silently failed to report a .startswith or .endswith call if either passed offset was -1 (the mechanism could not distinguish between that and an error)
 - Any trace_branch instrumentation added after fuzzing started (e.g. from lazily loaded modules) silently did not work.

PiperOrigin-RevId: 822737198
Change-Id: I7109a420907dff963b601a3879cefca77467dc54